rgw_file: store bucket uxattrs on the bucket
This turns out to be the corollary of the earlier bug in removing
buckets w/uxattrs. Rather than conditionally put the attrs on a '/'
object, put them on the bucket--because that works, and will be
slightly more efficient when the check for the '/' object goes
away.

Fixes: http://tracker.ceph.com/issues/20082
Resolves: rhbz#1450401

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit 448edbfa55b921c2cb58c4f051c8aeef15fe3c17)