osd: PrimaryLogPG, PGBackend: complete callback even if interval changes
This avoids leaking the callback passed to remove_missing_object().
Use EAGAIN to signal that the work was not completed, so we don't have
to take the pg lock and check if the interval changed again in
PGBackend.

Fixes: http://tracker.ceph.com/issues/20747
Signed-off-by: Josh Durgin <jdurgin@redhat.com>
(cherry picked from commit 3f591d4f43a455a273f334d82bab8c92aeb664a1)

Resolves: rhbz#1452780