RGW: S3 POST policy should not require Content-Type
In the current implementation, putting a key in the acting
RGWPolicyEnv makes it required in RGWPolicyEnv::match_policy_vars.

I'm not sure this is the intent, but in any case, add it to the env
only if sent.

Fixes: http://tracker.ceph.com/issues/20201
Resolves: rhbz#1502780

Signed-off-by: Matt Benjamin <mbenjamin@redhat.com>
(cherry picked from commit af29276c32f7009a78bd8b90d8f7d19026859c74)