rgw: ability to list bucket contents in unsorted order for efficiency
Adds the ability to list the contents of a bucket without the results
being sorted. This can have performance benefits since all bucket
shards do not need to be queried. This is done by adding a parameter
to the REST query string (i.e., "allow_unordered" [note underscore]
for swift or "allow-unordered" [note hyphen] for s3) to true.

This works with other request parameters such as marker, end_marker,
and prefix, but it cannot be used in conjunction with delimiter. Also,
because the results are not in order a later marker may precede an
earlier marker.

Resolves: rhbz#1595374

Original BZ: 1548563

Signed-off-by: J. Eric Ivancich <ivancich@redhat.com>